perform host healthcheck on startup
When the server is started, verify that each host in status up is still
reachable (that is, qarax-node is responding to a healthcheck).
As well as check hosts that were previously in unknown status.

Hosts will move to "initializing" status during the healthcheck and move
to either "up" or "unknown" depending on the result of the healthcheck.